pyproject.toml
[project]
name = "arlo-ai"
version = "0.1.0"
description = "Voice Assistant 'Arlo' - A powerful voice assistant that helps you control your computer and perform various tasks hands-free."
authors = [
    {name = "Dhruv",email = "you@example.com"}
]
license = {text = "EULA"}
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "aiohappyeyeballs (==2.4.4)",
    "aioconsole (==0.8.1)",
    "aiofiles (==24.1.0)",
    "aiohttp (==3.11.11)",
    "aiosignal (==1.3.2)",
    "attrs (==24.3.0)",
    "blingfire (==0.1.8)",
    "colorlog (==6.9.0)",
    "easyprocess (==1.1)",
    "edge-tts (==7.0.0)",
    "entrypoint2 (==1.1)",
    "frozenlist (==1.5.0)",
    "groq (==0.18.0)",
    "mss (==10.0.0)",
    "numpy (==2.2.1)",
    "opencv-python (==4.10.0.84)",
    "pillow (==11.1.0)",
    "propcache (==0.2.1)",
    "pvcobra (==2.0.5)",
    "pydantic (==2.10.5)",
    "pyperclip (==1.9.0)",
    "pyscreenshot (==3.1)",
    "python-dotenv (==1.0.1)",
    "sounddevice (==0.5.1)",
    "soundfile (==0.13.0)",
    "srt (==3.5.3)",
    "tabulate (==0.9.0)",
    "yarl (==1.18.3)",
    "pvporcupine (>=3.0.5,<4.0.0)",
    "faster-whisper (>=1.1.1,<2.0.0)",
    "validators (>=0.34.0,<0.35.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "langchain[groq] (>=0.3.20,<0.4.0)",
    "sentence-transformers (>=3.4.1,<4.0.0)",
    "chromadb (>=0.6.3,<0.7.0)",
    "langchain-community (>=0.3.20,<0.4.0)",
    "mediawikiapi (>=1.3,<2.0)",
    "wikibase-rest-api-client (>=0.2.2,<0.3.0)",
    "langgraph (>=0.3.21,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include ="src"}]
